#  DON'T MERGE: provider: keystore benchmark