Supersedes #1139 

> [!WARNING]
> Don't merge, this PR exists only for discussion + sharing the new keystore implementations + benchmark code

# KeyStore Benchmarking

Fixes: https://github.com/libp2p/go-libp2p-kad-dht/issues/1112

## TL;DR

From #1112, we suspected the provider `KeyStore` wasn't efficient on its operations, so I came up with different implementation to compare the performance. It turns out that a new implementation `KeyStore3` could be more optimized for the provider use case.

## Implementations

The keystore role is to store multihashes in a datastore, indexing them by their Kademlia identifier (256-bit digest, not included in the multihash).

The goal is to store the multihashes on a datastore to avoid having the mutlihashes of all CIDs the node is (re)providing in memory.

It must support:
* `Put([]multihash) []multihash`: add multihashes to the datastore, indexed by `kadid`. Return the newly added multihashes, but not the ones that were already stored.
* `Get(bitstringPrefix) []multihash`: get all the multihashes whose `kadid` matches the supplied bitstring prefix.
* `ContainsPrefix(bitstringPrefix) bool`: true if at least one multihash whose `kadid` matches the supplied bitstring prefix.
* `Delete([]multihash)`: remove multihashes from datastore 

`SweepingProvider` will call:
* `Put()` whenever keys are provided (`kubo` will mostly provide 1 key per call)
* `Get()` whenever it needs to reprovide a region (~every `24s` for the amino DHT, or `reprovide_interval / network_size * (1.5*replication_factor)`)
* `ContainsPrefix()`: once every `reprovide_interval` (`22h` in the amino DHT)
* `Delete()`: whenever `StopProviding()` is called by user (not used in Kubo).

### Old `KeyStore`

Strategy: store slices of multihashes whose `kadid` shares a common prefix of constant size.

* Initial implementation
* `Put([]multihash) []multihash`: 
  * Group multihashes by common prefix of constant size (given to constructor, `10` by default).
  * For each group, identified by a bitstring prefix, load existing multihashes matching the bitstring prefix
  * Combine all keys matching the bitstring prefix (new keys, and datastore keys), and write them back to the datastore (json marshal the slice).
* `Get(bitstringPrefix) []multihash`:
  * If prefix is shorter or equal to constant prefix size, load all (json) slices containing the multihashes whose `kadid` matches the supplied bitstring prefix. Return all of these multihashes
  * If prefix is longer than constant prefix size, load the slice that is a prefix, of the supplied bitstring prefix.
    * Out of the loaded multihashes, keep and return only the multihashes whose `kadid` matches the full bitstring prefix supplied.
* `ContainsPrefix(bitstringPrefix) bool`: very similar to `Get()`

The `Put()` operation seems unnatural, since we need to load, append and store, instead of simply storing.

### New `KeyStore2`

Strategy: store each multihash independently in the keystore, under the key `/base/prefix/1/0/1/1/...` with the binary part being the string representation of the `kadid` separated by slashes (to allow for prefix requests).

* `Put([]multihash) []multihash`: 
  * For each multihash, call `datastore.Has(multihash)`, if multihash is already stored skip this one 
  * store the multihash bytes directly under the key `/base/prefix/1/0/1/1/...` where the binary part is the string representation of the `kadid`
* `Get(bitstringPrefix) []multihash`:
  * Datastore prefix request with key `/base/prefix/0/0/` where `00` is the supplied prefix 
* `ContainsPrefix(bitstringPrefix) bool`: very similar to `Get()`
  * Datastore prefix request with limit of 1.

Super simple, but super long keys (>500 chars).

### New `KeyStore3`

Startegy: same as `KeyStore2` but use hex encoding instead of bitstring to optimize key size.

* `Put([]multihash) []multihash`: 
  * Same as `KeyStore2`, but the datastore key is `/base/prefix/4/f/1/3/...`
* `Get(bitstringPrefix) []multihash`:
  * Convert bitstringPrefix to hex (drop the last 1-3 bits if necessary)
  * if `len(bitstringPrefix)%4 != 0`, load the shorter prefix from datastore, and retrun only the multihashes matching the full bitstring prefix supplied 
* `ContainsPrefix(bitstringPrefix) bool`: very similar to `Get()`
  * Similar to `Get()`

Hybrid between `KeyStore` and `KeyStore2` (closer to `KeyStore2`)

## Benchmark

Here are a couple of benchmarks. You can find the benchmark code in the PR, and modify the parameters.

I tested against the following datastore implementations:
* `MapDatastore` (in memory)
* [`Badger`](https://github.com/ipfs/go-ds-badger4)
* [`LevelDB`](https://github.com/ipfs/go-ds-leveldb)
* [`Pebble`](https://github.com/ipfs/go-ds-pebble/)

### `Put` `1k` multihashes `100` times

<details>
  <summary>Results</summary>

```
➜  datastore git:(provider-keystore-benchmark) ✗ go test -bench=BenchmarkKeyStorePut -benchtime=3s -count=7 -cpu=1
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-libp2p-kad-dht/provider/datastore
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2154423625 ns/op	904194364 B/op	7541604 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2181195352 ns/op	904126260 B/op	7541569 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2060962842 ns/op	904126600 B/op	7541575 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2090892198 ns/op	904126936 B/op	7541583 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2110527342 ns/op	904127036 B/op	7541584 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2100944024 ns/op	904126692 B/op	7541577 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore         	      2	2095789098 ns/op	904126660 B/op	7541577 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	963107938 ns/op	346828048 B/op	1303191 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	964129954 ns/op	342362260 B/op	1302933 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      3	1020292245 ns/op	342362224 B/op	1302932 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	955377200 ns/op	342362240 B/op	1302932 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	957061942 ns/op	342362260 B/op	1302933 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	961950583 ns/op	342362264 B/op	1302933 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore2        	      4	945789228 ns/op	342362220 B/op	1302932 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	327878305 ns/op	128858075 B/op	1103025 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	324264951 ns/op	128858073 B/op	1103026 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	325651078 ns/op	128858068 B/op	1103026 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	324385938 ns/op	128858073 B/op	1103026 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	323993078 ns/op	128858073 B/op	1103026 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	317011156 ns/op	128868396 B/op	1103026 allocs/op
BenchmarkKeyStorePut/Map_1000_keys/KeyStore3        	     10	318319540 ns/op	128936774 B/op	1103029 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      2	2301118192 ns/op	1228887144 B/op	6420563 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	3870541996 ns/op	1724874728 B/op	10246227 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	3943736631 ns/op	1757120648 B/op	9705992 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	3883847922 ns/op	1721956688 B/op	10421963 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	4913411941 ns/op	2109809800 B/op	12118449 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	4689787947 ns/op	2107606536 B/op	10578895 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore2     	      1	3949379118 ns/op	2009892808 B/op	10576338 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      2	2027057832 ns/op	765009020 B/op	8156389 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      1	3119275640 ns/op	1084553456 B/op	9747691 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      2	2120541204 ns/op	660409900 B/op	9231093 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      2	2611539252 ns/op	702816484 B/op	10102265 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      1	3094105987 ns/op	950984528 B/op	10806480 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      1	3410545325 ns/op	849740720 B/op	11100441 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore3     	      1	3342187673 ns/op	716876168 B/op	11293270 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	5724569803 ns/op	3059094496 B/op	16637140 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	4688642886 ns/op	3219196136 B/op	13711436 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	3507623402 ns/op	2234517328 B/op	12177340 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	4660293656 ns/op	2869388616 B/op	14040614 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	3604241123 ns/op	2547720264 B/op	12569403 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	4784822004 ns/op	3188903568 B/op	14437698 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys/KeyStore      	      1	3688484742 ns/op	2547329832 B/op	12573392 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	3299493941 ns/op	1236382312 B/op	3735074 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4428536605 ns/op	1627709424 B/op	4775405 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4625790456 ns/op	1639590016 B/op	4831106 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4556883784 ns/op	1662939680 B/op	4838224 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4438684433 ns/op	1625544992 B/op	4866102 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4533176662 ns/op	1618636328 B/op	4857760 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore2    	      1	4430993578 ns/op	1608238480 B/op	4855259 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      2	1845614670 ns/op	516276372 B/op	3475430 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      3	1573051421 ns/op	377129976 B/op	3489742 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      2	1509441368 ns/op	364992416 B/op	3382469 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      3	1515945023 ns/op	373744192 B/op	3428941 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      2	1539081465 ns/op	371786492 B/op	3454858 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      2	1639632970 ns/op	384893164 B/op	3565454 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore3    	      2	1532659975 ns/op	366481232 B/op	3392315 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3792663029 ns/op	1339977160 B/op	9559356 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3227207509 ns/op	1277755680 B/op	8806708 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3161001322 ns/op	1286612200 B/op	8782474 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3201111016 ns/op	1281979448 B/op	8800547 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3246155900 ns/op	1275745584 B/op	8793346 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3278226364 ns/op	1274947344 B/op	8810259 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys/KeyStore     	      1	3391288239 ns/op	1261525824 B/op	8803568 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	4184442229 ns/op	596950224 B/op	1487375 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5156405579 ns/op	669754424 B/op	1723201 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5009835426 ns/op	668658992 B/op	1720997 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5027787819 ns/op	670987456 B/op	1732424 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5133910760 ns/op	670448320 B/op	1728701 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5062868452 ns/op	668447536 B/op	1724783 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore2     	      1	5182562843 ns/op	671824768 B/op	1730712 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2775514794 ns/op	221466168 B/op	1441242 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2696955158 ns/op	197484480 B/op	1429320 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2671293140 ns/op	196409188 B/op	1429338 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2602727800 ns/op	196236620 B/op	1427976 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2626474580 ns/op	196704820 B/op	1429305 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2669610064 ns/op	196535580 B/op	1430615 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore3     	      2	2906899378 ns/op	196067616 B/op	1429236 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4555627539 ns/op	1133656424 B/op	7953763 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4411314881 ns/op	1111270152 B/op	7744090 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4376105734 ns/op	1110658848 B/op	7744430 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4616323500 ns/op	1107766840 B/op	7744946 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4736400219 ns/op	1111327368 B/op	7744022 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4867563263 ns/op	1111046768 B/op	7744978 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys/KeyStore      	      1	4537264691 ns/op	1109939472 B/op	7744905 allocs/op
```

</details>

`KeyStore3` seems consistently better than the others.

### `Put` `1` multihash `100k` times

<details>
  <summary>Results</summary>

```
➜  datastore git:(provider-keystore-benchmark) ✗ go test -bench=BenchmarkKeyStorePut -benchtime=3s -count=7 -cpu=1
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-libp2p-kad-dht/provider/datastore
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4390708546 ns/op	1380856832 B/op	11802031 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4150009351 ns/op	1380721296 B/op	11801971 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4101984870 ns/op	1380721040 B/op	11801967 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4079597838 ns/op	1380720888 B/op	11801963 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4103068415 ns/op	1380721104 B/op	11801969 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4122328752 ns/op	1380720856 B/op	11801965 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore         	      1	4162538968 ns/op	1380721352 B/op	11801973 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1245533586 ns/op	370363200 B/op	1700362 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1291881087 ns/op	365195754 B/op	1700050 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1289523274 ns/op	364802266 B/op	1700033 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1286450269 ns/op	364802304 B/op	1700034 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1337495514 ns/op	364802250 B/op	1700032 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1306064786 ns/op	364802261 B/op	1700033 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore2        	      3	1294097971 ns/op	364802266 B/op	1700033 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	486403503 ns/op	152002107 B/op	1600027 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	475224263 ns/op	152002125 B/op	1600028 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	473031096 ns/op	152002118 B/op	1600028 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	482297342 ns/op	152002118 B/op	1600027 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	477492185 ns/op	152002125 B/op	1600028 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	477999655 ns/op	152002128 B/op	1600028 allocs/op
BenchmarkKeyStorePut/Map_1_keys/KeyStore3        	      7	635974555 ns/op	152002139 B/op	1600028 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	8543308963 ns/op	3170723936 B/op	18166763 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	9840355250 ns/op	4156397216 B/op	20566727 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	9756807332 ns/op	3828224344 B/op	20954564 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	9454734042 ns/op	4240293912 B/op	20687285 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	8483858629 ns/op	3842193376 B/op	19567497 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	9393398147 ns/op	3873776264 B/op	20807481 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore      	      1	11849024528 ns/op	4629808224 B/op	21083619 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	3204046153 ns/op	1140847728 B/op	8262652 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	6859639393 ns/op	2110335280 B/op	14156963 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	6261886275 ns/op	1693530680 B/op	13773045 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	7843947485 ns/op	1984901192 B/op	14934634 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	9085735320 ns/op	2571216024 B/op	15303545 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	9080651908 ns/op	2029152472 B/op	14132010 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore2     	      1	7838518900 ns/op	2109720896 B/op	15276737 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	5421562048 ns/op	1546935656 B/op	13202917 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	3494783829 ns/op	726254128 B/op	12324715 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	4358501882 ns/op	1014130296 B/op	13255119 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	4850065606 ns/op	915281552 B/op	13539894 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	5050315038 ns/op	788094712 B/op	13753669 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	5982616135 ns/op	1118635624 B/op	14514864 allocs/op
BenchmarkKeyStorePut/Badger_1_keys/KeyStore3     	      1	6562412176 ns/op	1123572216 B/op	14676061 allocs/op
BenchmarkKeyStorePut/LevelDB_1_keys/KeyStore     	      1	7079311500 ns/op	2141169552 B/op	13693567 allocs/op
BenchmarkKeyStorePut/LevelDB_1_keys/KeyStore     	      1	9370605318 ns/op	2110908504 B/op	13712062 allocs/op
BenchmarkKeyStorePut/LevelDB_1_keys/KeyStore     	      1	7059329813 ns/op	2123407536 B/op	13767253 allocs/op
…