@guillaumemichel 2025-09-05
Added a `KeyStore4` implementations with same performance level (slightly better) than `KeyStore3` that was the best so far.

<details>
  <summary>Results</summary>

```
➜  datastore git:(provider-keystore-benchmark-2) ✗ go test -bench=BenchmarkKeyStorePut -benchtime=3s -count=4 -cpu=1 -timeout 60m
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-libp2p-kad-dht/provider/datastore
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore2         	      2	2371162956 ns/op	1232181984 B/op	6439166 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore2         	      1	3911365960 ns/op	1716305584 B/op	10227951 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore2         	      1	4319190002 ns/op	1846744656 B/op	9710823 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore2         	      1	3742240655 ns/op	1604041976 B/op	10421117 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore3         	      2	2316344664 ns/op	836645488 B/op	8410622 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore3         	      1	3541372359 ns/op	1102910040 B/op	9728034 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore3         	      2	2314933536 ns/op	686914556 B/op	8907303 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore3         	      2	2775347808 ns/op	673591032 B/op	9687981 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore4         	      1	3254648807 ns/op	810643000 B/op	10555039 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore4         	      1	3308536446 ns/op	657145296 B/op	10511179 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore4         	      1	3833692962 ns/op	997803528 B/op	11287183 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore4         	      1	7314594744 ns/op	1085526352 B/op	11687466 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore          	      1	5489089390 ns/op	2614460032 B/op	16033823 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore          	      1	5756408624 ns/op	2627945432 B/op	13715375 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore          	      1	4110173496 ns/op	2488636624 B/op	12300187 allocs/op
BenchmarkKeyStorePut/Badger_1000_keys_100x/KeyStore          	      1	4679148895 ns/op	2774960616 B/op	14091491 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore         	      1	4538698731 ns/op	1621225560 B/op	9555257 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore         	      1	3695286515 ns/op	1318794344 B/op	8857989 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore         	      1	3642305164 ns/op	1322969944 B/op	8892430 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore         	      1	3652753638 ns/op	1335763704 B/op	8900898 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore2        	      1	3531583049 ns/op	1273664680 B/op	4033789 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore2        	      1	4465996543 ns/op	1599048992 B/op	4826441 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore2        	      1	4440514717 ns/op	1629375048 B/op	4926617 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore2        	      1	4528793187 ns/op	1651111952 B/op	4950462 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore3        	      2	1811280898 ns/op	522256944 B/op	3507743 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore3        	      2	1527815714 ns/op	372191848 B/op	3457299 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore3        	      3	1457702673 ns/op	374817202 B/op	3456920 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore3        	      3	1438774519 ns/op	370360514 B/op	3455516 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore4        	      2	1555510924 ns/op	383970900 B/op	3567544 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore4        	      3	1488687114 ns/op	376389069 B/op	3467723 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore4        	      2	1518305506 ns/op	382325564 B/op	3567047 allocs/op
BenchmarkKeyStorePut/LevelDB_1000_keys_100x/KeyStore4        	      3	1544172105 ns/op	382605181 B/op	3530454 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore          	      1	3177993481 ns/op	1108391736 B/op	7739426 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore          	      1	3303318278 ns/op	1111951840 B/op	7742556 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore          	      1	3204445610 ns/op	1115305720 B/op	7743765 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore          	      1	3235368854 ns/op	1113859952 B/op	7743950 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore2         	      1	3991893886 ns/op	599521808 B/op	1481178 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore2         	      1	4879270856 ns/op	669078856 B/op	1721712 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore2         	      1	5281542853 ns/op	669651736 B/op	1730655 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore2         	      1	4989274813 ns/op	671792168 B/op	1730385 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore3         	      1	3046548037 ns/op	245898032 B/op	1462471 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore3         	      2	2327865223 ns/op	196104252 B/op	1428619 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore3         	      2	2253748062 ns/op	196916292 B/op	1427898 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore3         	      2	2318348462 ns/op	196556000 B/op	1428679 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore4         	      2	2277493135 ns/op	196083420 B/op	1425703 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore4         	      2	2233418511 ns/op	196359504 B/op	1426087 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore4         	      2	2150149736 ns/op	196898076 B/op	1426502 allocs/op
BenchmarkKeyStorePut/Pebble_1000_keys_100x/KeyStore4         	      2	2218946998 ns/op	196386376 B/op	1426831 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-kad-dht/provider/datastore	207.175s
```

</details>